added fs and pdf generation

+++ controllers/admin.js
@@ -1,6 +1,8 @@
 const mongodb = require('mongodb');
 const Product = require('../models/product');
 const { validationResult } = require('express-validator');
+const product = require('../models/product');
+const fileHelper = require('../util/file');
 
 //GET
 exports.getProducts = (req, res) => {
@@ -49,21 +51,32 @@ exports.getAddProduct = (req, res) => {
 //POST
 exports.postDeleteProduct = (req, res) => {
   const prodId = req.body.productId;
-  Product.deleteOne({ _id: prodId, userId: req.user._id })
+  Product.findById(prodId)
+    .then((product) => {
+      if (!product) {
+        return next(new Error('Product not found'));
+      }
+      fileHelper.deleteFile(product.imageUrl);
+      return Product.deleteOne({ _id: prodId, userId: req.user._id });
+    })
     .then(() => {
       console.log('deleted product');
       res.redirect('/admin/products');
     })
-    .catch((err) => console.log(err));
+    .catch((err) => {
+      const error = new Error(err);
+      error.httpStatusCode = 500;
+      return next(error);
+    });
 };
 exports.postEditProduct = (req, res) => {
   const prodId = req.body.productId,
     updatedTitle = req.body.title,
     updatedPrice = req.body.price,
     updatedDesc = req.body.description,
-    updatedUrl = req.body.imageUrl;
+    image = req.file;
   const errors = validationResult(req);
-  console.log(errors.mapped());
+
   if (!errors.isEmpty()) {
     return res.status(422).render('admin/editProduct', {
       docTitle: 'edit product',
@@ -73,7 +86,7 @@ exports.postEditProduct = (req, res) => {
       product: {
         _id: prodId,
         title: updatedTitle,
-        imageUrl: updatedUrl,
+        imageUrl: image.path,
         price: updatedPrice,
         description: updatedDesc,
       },
@@ -87,7 +100,10 @@ exports.postEditProduct = (req, res) => {
         return res.redirect('/');
       }
       product.title = updatedTitle;
-      product.imageUrl = updatedUrl;
+      if (image) {
+        fileHelper.deleteFile(product.imageUrl);
+        product.imageUrl = image.path;
+      }
       product.price = updatedPrice;
       product.description = updatedDesc;
       return product.save().then(() => {
@@ -95,18 +111,36 @@ exports.postEditProduct = (req, res) => {
         res.redirect('/admin/products');
       });
     })
-    .catch((err) => console.log(err));
+    .catch((err) => {
+      const error = new Error(err);
+      error.httpStatusCode = 500;
+      return next(error);
+    });
 };
 exports.postAddProduct = (req, res, next) => {
   const title = req.body.title;
-  let imageUrl = req.body.imageUrl;
+  let image = req.file;
   const price = req.body.price;
   const description = req.body.description;
   const errors = validationResult(req);
 
-  console.log(errors.array());
+  if (!image) {
+    return res.status(422).render('admin/editProduct', {
+      docTitle: 'add product',
+      path: '/admin/add-products',
+      editing: false,
+      hasError: true,
+      product: {
+        title,
+        price,
+        description,
+      },
+      errorMessage: 'Attached file is not an image',
+      validationErrors: [],
+    });
+  }
   if (!errors.isEmpty()) {
-    return res.status(422).render('admin/addProduct', {
+    return res.status(422).render('admin/editProduct', {
       docTitle: 'add product',
       path: '/admin/add-products',
       editing: false,
@@ -120,15 +154,12 @@ exports.postAddProduct = (req, res, next) => {
       validationErrors: errors.mapped(),
     });
   }
+  const imageUrl = image.path;
   const product = new Product({
-    title: title,
-    price: price,
-    imageUrl:
-      imageUrl ||
-      `https://picsum.photos/seed/${encodeURIComponent(
-        title.toLowerCase().trim()
-      )}/400/600`,
-    description: description,
+    title,
+    price,
+    imageUrl: imageUrl,
+    description,
     userId: req.session.user,
   });
   product
@@ -137,7 +168,11 @@ exports.postAddProduct = (req, res, next) => {
       console.log('Created Product');
       res.redirect('/admin/products');
     })
-    .catch((err) => console.log(err));
+    .catch((err) => {
+      const error = new Error(err);
+      error.httpStatusCode = 500;
+      return next(error);
+    });
 };
